refactor: convert indexer infra tests to dango test setup, retire `TestBuilder`
Convert `indexer_infra/sql.rs` and `hooked.rs` from using
`TestBuilder` + `create_hooked_indexer()` to `setup_test_naive_with_indexer`.
This retires `TestBuilder` entirely from the `dango-testing` crate.

Key changes:
- Remove `dango/testing/src/indexer/setup.rs` (and its `create_hooked_indexer`)
- Inline `indexer/request.rs` into `indexer.rs` (single-file module)
- Add `indexer_cache::context::Context` to the setup return tuples
- Update all callers for the expanded return signature
