+++ dango/testing/tests/httpd/accounts.rs
@@ -25,8 +25,17 @@ use {
 
 #[tokio::test(flavor = "multi_thread")]
 async fn query_accounts() -> anyhow::Result<()> {
-    let (suite, mut accounts, codes, contracts, validator_sets, dango_httpd_context, _, _db_guard) =
-        setup_test_with_indexer(TestOption::default()).await;
+    let (
+        suite,
+        mut accounts,
+        codes,
+        contracts,
+        validator_sets,
+        dango_httpd_context,
+        _,
+        _,
+        _db_guard,
+    ) = setup_test_with_indexer(TestOption::default()).await;
     let mut suite = HyperlaneTestSuite::new(suite, validator_sets, &contracts);
 
     let user1 = create_user_and_account(&mut suite, &mut accounts, &contracts, &codes).await;
@@ -63,8 +72,17 @@ async fn query_accounts() -> anyhow::Result<()> {
 
 #[tokio::test(flavor = "multi_thread")]
 async fn query_accounts_with_user_index() -> anyhow::Result<()> {
-    let (suite, mut accounts, codes, contracts, validator_sets, dango_httpd_context, _, _db_guard) =
-        setup_test_with_indexer(TestOption::default()).await;
+    let (
+        suite,
+        mut accounts,
+        codes,
+        contracts,
+        validator_sets,
+        dango_httpd_context,
+        _,
+        _,
+        _db_guard,
+    ) = setup_test_with_indexer(TestOption::default()).await;
     let mut suite = HyperlaneTestSuite::new(suite, validator_sets, &contracts);
 
     let user = create_user_and_account(&mut suite, &mut accounts, &contracts, &codes).await;
@@ -106,8 +124,17 @@ async fn query_accounts_with_user_index() -> anyhow::Result<()> {
 
 #[tokio::test(flavor = "multi_thread")]
 async fn query_accounts_with_wrong_user_index() -> anyhow::Result<()> {
-    let (suite, mut accounts, codes, contracts, validator_sets, dango_httpd_context, _, _db_guard) =
-        setup_test_with_indexer(TestOption::default()).await;
+    let (
+        suite,
+        mut accounts,
+        codes,
+        contracts,
+        validator_sets,
+        dango_httpd_context,
+        _,
+        _,
+        _db_guard,
+    ) = setup_test_with_indexer(TestOption::default()).await;
     let mut suite = HyperlaneTestSuite::new(suite, validator_sets, &contracts);
 
     create_user_and_account(&mut suite, &mut accounts, &contracts, &codes).await;
@@ -145,8 +172,17 @@ async fn query_accounts_with_wrong_user_index() -> anyhow::Result<()> {
 #[ignore = "flaky"]
 #[tokio::test(flavor = "multi_thread")]
 async fn query_user_multiple_single_signature_accounts() -> anyhow::Result<()> {
-    let (suite, mut accounts, codes, contracts, validator_sets, dango_httpd_context, _, _db_guard) =
-        setup_test_with_indexer(TestOption::default()).await;
+    let (
+        suite,
+        mut accounts,
+        codes,
+        contracts,
+        validator_sets,
+        dango_httpd_context,
+        _,
+        _,
+        _db_guard,
+    ) = setup_test_with_indexer(TestOption::default()).await;
     let mut suite = HyperlaneTestSuite::new(suite, validator_sets, &contracts);
 
     // Create two accounts under the same user. The two `TestAccount`'s should
@@ -227,8 +263,17 @@ async fn query_user_multiple_single_signature_accounts() -> anyhow::Result<()> {
 
 #[tokio::test(flavor = "multi_thread")]
 async fn graphql_paginate_accounts() -> anyhow::Result<()> {
-    let (suite, mut accounts, codes, contracts, validator_sets, dango_httpd_context, _, _db_guard) =
-        setup_test_with_indexer(TestOption::default()).await;
+    let (
+        suite,
+        mut accounts,
+        codes,
+        contracts,
+        validator_sets,
+        dango_httpd_context,
+        _,
+        _,
+        _db_guard,
+    ) = setup_test_with_indexer(TestOption::default()).await;
     let mut suite = HyperlaneTestSuite::new(suite, validator_sets, &contracts);
 
     // Create 10 accounts to paginate through
@@ -332,8 +377,17 @@ async fn graphql_paginate_accounts() -> anyhow::Result<()> {
 
 #[tokio::test(flavor = "multi_thread")]
 async fn graphql_subscribe_to_accounts() -> anyhow::Result<()> {
-    let (suite, mut accounts, codes, contracts, validator_sets, dango_httpd_context, _, _db_guard) =
-        setup_test_with_indexer(TestOption::default()).await;
+    let (
+        suite,
+        mut accounts,
+        codes,
+        contracts,
+        validator_sets,
+        dango_httpd_context,
+        _,
+        _,
+        _db_guard,
+    ) = setup_test_with_indexer(TestOption::default()).await;
     let mut suite = HyperlaneTestSuite::new(suite, validator_sets, &contracts);
 
     let _test_account =
@@ -408,8 +462,17 @@ async fn graphql_subscribe_to_accounts() -> anyhow::Result<()> {
 
 #[tokio::test(flavor = "multi_thread")]
 async fn graphql_subscribe_to_accounts_with_user_index() -> anyhow::Result<()> {
-    let (suite, mut accounts, codes, contracts, validator_sets, dango_httpd_context, _, _db_guard) =
-        setup_test_with_indexer(TestOption::default()).await;
+    let (
+        suite,
+        mut accounts,
+        codes,
+        contracts,
+        validator_sets,
+        dango_httpd_context,
+        _,
+        _,
+        _db_guard,
+    ) = setup_test_with_indexer(TestOption::default()).await;
     let mut suite = HyperlaneTestSuite::new(suite, validator_sets, &contracts);
 
     let mut test_account1 =
@@ -484,7 +547,7 @@ async fn graphql_subscribe_to_accounts_with_user_index() -> anyhow::Result<()> {
 
 #[tokio::test(flavor = "multi_thread")]
 async fn graphql_returns_account_owner_nonces() -> anyhow::Result<()> {
-    let (mut suite, mut accounts, _, _, _, dango_httpd_context, _, _db_guard) =
+    let (mut suite, mut accounts, _, _, _, dango_httpd_context, _, _, _db_guard) =
         setup_test_with_indexer(TestOption::default()).await;
 
     // copied from `tracked_nonces_works``
@@ -545,7 +608,7 @@ async fn graphql_returns_account_owner_nonces() -> anyhow::Result<()> {
 
 #[tokio::test(flavor = "multi_thread")]
 async fn graphql_returns_address_balance() -> anyhow::Result<()> {
-    let (mut suite, mut accounts, _, _, _, dango_httpd_context, _, _db_guard) =
+    let (mut suite, mut accounts, _, _, _, dango_httpd_context, _, _, _db_guard) =
         setup_test_with_indexer(TestOption::default()).await;
 
     // copied from `tracked_nonces_works``
