+++ dango/testing/tests/httpd/perps_events.rs
@@ -15,7 +15,7 @@ use {
 /// are indexed after a limit-order match.
 #[tokio::test(flavor = "multi_thread")]
 async fn query_perps_events_user_lifecycle() -> anyhow::Result<()> {
-    let (mut suite, mut accounts, _, contracts, _, dango_httpd_context, _, _db_guard) =
+    let (mut suite, mut accounts, _, contracts, _, dango_httpd_context, _, _, _db_guard) =
         setup_test_with_indexer(TestOption::default()).await;
 
     let pair = pair_id();
