+++ dango/testing/tests/httpd/perps_pair_stats.rs
@@ -14,7 +14,7 @@ use {
 
 #[tokio::test(flavor = "multi_thread")]
 async fn query_perps_pair_stats() -> anyhow::Result<()> {
-    let (mut suite, mut accounts, _, contracts, _, dango_httpd_context, _, _db_guard) =
+    let (mut suite, mut accounts, _, contracts, _, dango_httpd_context, _, _, _db_guard) =
         setup_test_with_indexer(TestOption::default()).await;
 
     let pair = pair_id();
@@ -72,7 +72,7 @@ async fn query_perps_pair_stats() -> anyhow::Result<()> {
 
 #[tokio::test(flavor = "multi_thread")]
 async fn query_perps_pair_stats_nonexistent_pair() -> anyhow::Result<()> {
-    let (suite, _, _, _, _, dango_httpd_context, _, _db_guard) =
+    let (suite, _, _, _, _, dango_httpd_context, _, _, _db_guard) =
         setup_test_with_indexer(TestOption::default()).await;
 
     suite.app.indexer.wait_for_finish().await?;
@@ -110,7 +110,7 @@ async fn query_perps_pair_stats_nonexistent_pair() -> anyhow::Result<()> {
 
 #[tokio::test(flavor = "multi_thread")]
 async fn query_all_perps_pair_stats() -> anyhow::Result<()> {
-    let (mut suite, mut accounts, _, contracts, _, dango_httpd_context, _, _db_guard) =
+    let (mut suite, mut accounts, _, contracts, _, dango_httpd_context, _, _, _db_guard) =
         setup_test_with_indexer(TestOption::default()).await;
 
     let pair = pair_id();
@@ -162,7 +162,7 @@ async fn query_all_perps_pair_stats() -> anyhow::Result<()> {
 
 #[tokio::test(flavor = "multi_thread")]
 async fn query_perps_pair_stats_partial_fields() -> anyhow::Result<()> {
-    let (mut suite, mut accounts, _, contracts, _, dango_httpd_context, _, _db_guard) =
+    let (mut suite, mut accounts, _, contracts, _, dango_httpd_context, _, _, _db_guard) =
         setup_test_with_indexer(TestOption::default()).await;
 
     let pair = pair_id();
