+++ dango/testing/tests/httpd/transfers.rs
@@ -17,7 +17,7 @@ use {
 
 #[tokio::test(flavor = "multi_thread")]
 async fn graphql_returns_transfer_and_accounts() -> anyhow::Result<()> {
-    let (mut suite, mut accounts, _, contracts, _, dango_httpd_context, _, _db_guard) =
+    let (mut suite, mut accounts, _, contracts, _, dango_httpd_context, _, _, _db_guard) =
         setup_test_with_indexer(TestOption::default()).await;
 
     // Copied from benchmarks.rs
@@ -93,8 +93,17 @@ async fn graphql_returns_transfer_and_accounts() -> anyhow::Result<()> {
 
 #[tokio::test(flavor = "multi_thread")]
 async fn graphql_transfers_with_user_index() -> anyhow::Result<()> {
-    let (suite, mut accounts, codes, contracts, validator_sets, dango_httpd_context, _, _db_guard) =
-        setup_test_with_indexer(TestOption::default()).await;
+    let (
+        suite,
+        mut accounts,
+        codes,
+        contracts,
+        validator_sets,
+        dango_httpd_context,
+        _,
+        _,
+        _db_guard,
+    ) = setup_test_with_indexer(TestOption::default()).await;
 
     let mut suite = HyperlaneTestSuite::new(suite, validator_sets, &contracts);
 
@@ -192,8 +201,17 @@ async fn graphql_transfers_with_user_index() -> anyhow::Result<()> {
 
 #[tokio::test(flavor = "multi_thread")]
 async fn graphql_transfers_with_wrong_user_index() -> anyhow::Result<()> {
-    let (suite, mut accounts, codes, contracts, validator_sets, dango_httpd_context, _, _db_guard) =
-        setup_test_with_indexer(TestOption::default()).await;
+    let (
+        suite,
+        mut accounts,
+        codes,
+        contracts,
+        validator_sets,
+        dango_httpd_context,
+        _,
+        _,
+        _db_guard,
+    ) = setup_test_with_indexer(TestOption::default()).await;
 
     let mut suite = HyperlaneTestSuite::new(suite, validator_sets, &contracts);
 
@@ -239,7 +257,7 @@ async fn graphql_transfers_with_wrong_user_index() -> anyhow::Result<()> {
 
 #[tokio::test(flavor = "multi_thread")]
 async fn graphql_paginate_transfers() -> anyhow::Result<()> {
-    let (mut suite, mut accounts, _, _, _, dango_httpd_context, _, _db_guard) =
+    let (mut suite, mut accounts, _, _, _, dango_httpd_context, _, _, _db_guard) =
         setup_test_with_indexer(TestOption::default()).await;
 
     // Create 10 transfers to paginate through
@@ -355,7 +373,7 @@ async fn graphql_paginate_transfers() -> anyhow::Result<()> {
 
 #[tokio::test(flavor = "multi_thread")]
 async fn graphql_subscribe_to_transfers() -> anyhow::Result<()> {
-    let (mut suite, mut accounts, _, contracts, _, dango_httpd_context, _, _db_guard) =
+    let (mut suite, mut accounts, _, contracts, _, dango_httpd_context, _, _, _db_guard) =
         setup_test_with_indexer(TestOption::default()).await;
 
     // Copied from benchmarks.rs
@@ -471,7 +489,7 @@ async fn graphql_subscribe_to_transfers() -> anyhow::Result<()> {
 
 #[tokio::test(flavor = "multi_thread")]
 async fn graphql_subscribe_to_transfers_with_filter() -> anyhow::Result<()> {
-    let (mut suite, mut accounts, _, contracts, _, dango_httpd_context, _, _db_guard) =
+    let (mut suite, mut accounts, _, contracts, _, dango_httpd_context, _, _, _db_guard) =
         setup_test_with_indexer(TestOption::default()).await;
 
     // Copied from benchmarks.rs
