+++ dango/testing/tests/httpd/users.rs
@@ -11,8 +11,17 @@ use {
 
 #[tokio::test(flavor = "multi_thread")]
 async fn query_user() -> anyhow::Result<()> {
-    let (suite, mut accounts, codes, contracts, validator_sets, dango_httpd_context, _, _db_guard) =
-        setup_test_with_indexer(TestOption::default()).await;
+    let (
+        suite,
+        mut accounts,
+        codes,
+        contracts,
+        validator_sets,
+        dango_httpd_context,
+        _,
+        _,
+        _db_guard,
+    ) = setup_test_with_indexer(TestOption::default()).await;
     let mut suite = HyperlaneTestSuite::new(suite, validator_sets, &contracts);
 
     let user = create_user_and_account(&mut suite, &mut accounts, &contracts, &codes).await;
@@ -52,8 +61,17 @@ async fn query_user() -> anyhow::Result<()> {
 
 #[tokio::test(flavor = "multi_thread")]
 async fn query_single_user_multiple_public_keys() -> anyhow::Result<()> {
-    let (suite, mut accounts, codes, contracts, validator_sets, dango_httpd_context, _, _db_guard) =
-        setup_test_with_indexer(TestOption::default()).await;
+    let (
+        suite,
+        mut accounts,
+        codes,
+        contracts,
+        validator_sets,
+        dango_httpd_context,
+        _,
+        _,
+        _db_guard,
+    ) = setup_test_with_indexer(TestOption::default()).await;
     let mut suite = HyperlaneTestSuite::new(suite, validator_sets, &contracts);
 
     let mut test_account =
@@ -106,8 +124,17 @@ async fn query_single_user_multiple_public_keys() -> anyhow::Result<()> {
 
 #[tokio::test(flavor = "multi_thread")]
 async fn query_public_keys_by_user_index() -> anyhow::Result<()> {
-    let (suite, mut accounts, codes, contracts, validator_sets, dango_httpd_context, _, _db_guard) =
-        setup_test_with_indexer(TestOption::default()).await;
+    let (
+        suite,
+        mut accounts,
+        codes,
+        contracts,
+        validator_sets,
+        dango_httpd_context,
+        _,
+        _,
+        _db_guard,
+    ) = setup_test_with_indexer(TestOption::default()).await;
     let mut suite = HyperlaneTestSuite::new(suite, validator_sets, &contracts);
 
     let test_account = create_user_and_account(&mut suite, &mut accounts, &contracts, &codes).await;
@@ -149,8 +176,17 @@ async fn query_public_keys_by_user_index() -> anyhow::Result<()> {
 
 #[tokio::test(flavor = "multi_thread")]
 async fn query_users_rejects_conflicting_pagination_args() -> anyhow::Result<()> {
-    let (_suite, _accounts, _codes, _contracts, _validator_sets, dango_httpd_context, _, _db_guard) =
-        setup_test_with_indexer(TestOption::default()).await;
+    let (
+        _suite,
+        _accounts,
+        _codes,
+        _contracts,
+        _validator_sets,
+        dango_httpd_context,
+        _,
+        _,
+        _db_guard,
+    ) = setup_test_with_indexer(TestOption::default()).await;
 
     let local_set = tokio::task::LocalSet::new();
