+++ dango/testing/tests/indexer_infra/hooked.rs
@@ -1,12 +1,11 @@
 use {
     crate::sql::replier,
     assertor::*,
-    dango_testing::create_hooked_indexer,
-    dango_types::config::AppConfig,
+    dango_testing::{TestOption, setup_test_naive_with_indexer},
+    dango_types::constants::usdc,
     grug_app::{Db, Indexer},
-    grug_testing::TestBuilder,
     grug_types::{
-        Block, BlockInfo, BlockOutcome, Coin, Coins, Denom, Empty, Hash, Message, ReplyOn,
+        Addressable, Block, BlockInfo, BlockOutcome, Coins, Empty, Hash, Message, ReplyOn,
         ResultExt,
     },
     grug_vm_rust::ContractBuilder,
@@ -14,28 +13,20 @@ use {
     indexer_sql::entity,
     replier::{ExecuteMsg, ReplyMsg},
     sea_orm::{ColumnTrait, EntityTrait, QueryFilter, QueryOrder},
-    std::str::FromStr,
 };
 
 #[tokio::test(flavor = "multi_thread")]
 async fn index_block() {
-    let denom = Denom::from_str("ugrug").unwrap();
+    let (mut suite, mut accounts, _, _, _, httpd_context, _, _, _db_guard) =
+        setup_test_naive_with_indexer(TestOption::default()).await;
 
-    let (hooked_indexer, indexer_context, _) = create_hooked_indexer().await;
-
-    let (mut suite, mut accounts) = TestBuilder::new_with_indexer(hooked_indexer)
-        .add_account("owner", Coins::new())
-        .add_account("sender", Coins::one(denom.clone(), 30_000).unwrap())
-        .set_owner("owner")
-        .build();
-
-    let to = accounts["owner"].address;
+    let to = accounts.owner.address();
 
     suite
         .send_message_with_gas(
-            &mut accounts["sender"],
-            2000,
-            Message::transfer(to, Coins::one(denom.clone(), 2_000).unwrap()).unwrap(),
+            &mut accounts.user1,
+            1_000_000,
+            Message::transfer(to, Coins::one(usdc::DENOM.clone(), 100).unwrap()).unwrap(),
         )
         .await
         .should_succeed();
@@ -50,27 +41,27 @@ async fn index_block() {
 
     // ensure block was saved
     let block = entity::blocks::Entity::find()
-        .one(&indexer_context.db)
+        .one(&httpd_context.db)
         .await
         .expect("Can't fetch blocks");
     assert_that!(block).is_some();
     assert_that!(block.unwrap().block_height).is_equal_to(1);
 
     let transactions = entity::transactions::Entity::find()
-        .all(&indexer_context.db)
+        .all(&httpd_context.db)
         .await
         .expect("Can't fetch transactions");
     assert_that!(transactions).is_not_empty();
 
     let messages = entity::messages::Entity::find()
-        .all(&indexer_context.db)
+        .all(&httpd_context.db)
         .await
         .expect("Can't fetch messages");
 
     assert_that!(messages).is_not_empty();
 
     let events = entity::events::Entity::find()
-        .all(&indexer_context.db)
+        .all(&httpd_context.db)
         .await
         .expect("Can't fetch events");
 
@@ -87,25 +78,16 @@ async fn index_block() {
 /// before it was indexed.
 #[tokio::test(flavor = "multi_thread")]
 async fn parse_previous_block_after_restart() {
-    let denom = Denom::from_str("ugrug").unwrap();
+    let (mut suite, mut accounts, _, _, _, httpd_context, _, _, _db_guard) =
+        setup_test_naive_with_indexer(TestOption::default()).await;
 
-    let (indexer, indexer_context, _) = create_hooked_indexer().await;
-
-    let (mut suite, mut accounts) = TestBuilder::new_with_indexer(indexer)
-        .set_app_config(&AppConfig::default())
-        .unwrap()
-        .add_account("owner", Coins::new())
-        .add_account("sender", Coins::one(denom.clone(), 30_000).unwrap())
-        .set_owner("owner")
-        .build();
-
-    let to = accounts["owner"].address;
+    let to = accounts.owner.address();
 
     suite
         .send_message_with_gas(
-            &mut accounts["sender"],
-            2000,
-            Message::transfer(to, Coins::one(denom.clone(), 2_000).unwrap()).unwrap(),
+            &mut accounts.user1,
+            1_000_000,
+            Message::transfer(to, Coins::one(usdc::DENOM.clone(), 100).unwrap()).unwrap(),
         )
         .await
         .should_succeed();
@@ -128,13 +110,13 @@ async fn parse_previous_block_after_restart() {
     tracing::warn!("Shut down indexer");
 
     // 1. Delete database block height 1
-    entity::blocks::Entity::delete_block_and_data(&indexer_context.db, 1)
+    entity::blocks::Entity::delete_block_and_data(&httpd_context.db, 1)
         .await
         .unwrap();
 
     // 1 bis. Verify the block height 1 is deleted
     let block = entity::blocks::Entity::find()
-        .one(&indexer_context.db)
+        .one(&httpd_context.db)
         .await
         .expect("Can't fetch blocks");
     assert_that!(block).is_none();
@@ -153,7 +135,7 @@ async fn parse_previous_block_after_restart() {
 
     // 4. Verify the block height 1 is indexed
     let block = entity::blocks::Entity::find()
-        .one(&indexer_context.db)
+        .one(&httpd_context.db)
         .await
         .expect("Can't fetch blocks");
     assert_that!(block).is_some();
@@ -162,9 +144,9 @@ async fn parse_previous_block_after_restart() {
     // 5. Send a transaction
     suite
         .send_message_with_gas(
-            &mut accounts["sender"],
-            2000,
-            Message::transfer(to, Coins::one(denom.clone(), 2_000).unwrap()).unwrap(),
+            &mut accounts.user1,
+            1_000_000,
+            Message::transfer(to, Coins::one(usdc::DENOM.clone(), 100).unwrap()).unwrap(),
         )
         .await
         .should_succeed();
@@ -180,7 +162,7 @@ async fn parse_previous_block_after_restart() {
     // 6. Verify the block height 2 is indexed
     let block = entity::blocks::Entity::find()
         .order_by_desc(entity::blocks::Column::BlockHeight)
-        .one(&indexer_context.db)
+        .one(&httpd_context.db)
         .await
         .expect("Can't fetch blocks");
     assert_that!(block).is_some();
@@ -192,23 +174,16 @@ async fn parse_previous_block_after_restart() {
 /// after it was indexed, and before the tmp file was deleted.
 #[tokio::test(flavor = "multi_thread")]
 async fn no_sql_index_error_after_restart() {
-    let denom = Denom::from_str("ugrug").unwrap();
+    let (mut suite, mut accounts, _, _, _, httpd_context, cache_context, _, _db_guard) =
+        setup_test_naive_with_indexer(TestOption::default()).await;
 
-    let (indexer, sql_indexer_context, cache_context) = create_hooked_indexer().await;
-
-    let (mut suite, mut accounts) = TestBuilder::new_with_indexer(indexer)
-        .add_account("owner", Coins::new())
-        .add_account("sender", Coins::one(denom.clone(), 30_000).unwrap())
-        .set_owner("owner")
-        .build();
-
-    let to = accounts["owner"].address;
+    let to = accounts.owner.address();
 
     suite
         .send_message_with_gas(
-            &mut accounts["sender"],
-            2000,
-            Message::transfer(to, Coins::one(denom.clone(), 2_000).unwrap()).unwrap(),
+            &mut accounts.user1,
+            1_000_000,
+            Message::transfer(to, Coins::one(usdc::DENOM.clone(), 100).unwrap()).unwrap(),
         )
         .await
         .should_succeed();
@@ -230,7 +205,7 @@ async fn no_sql_index_error_after_restart() {
 
     // 1. Verify the block height 1 is indexed
     let block = entity::blocks::Entity::find()
-        .one(&sql_indexer_context.db)
+        .one(&httpd_context.db)
         .await
         .expect("Can't fetch blocks");
     assert_that!(block).is_some();
@@ -271,7 +246,7 @@ async fn no_sql_index_error_after_restart() {
 
     // 4. Verify the block height 1 is still indexed
     let block = entity::blocks::Entity::find()
-        .one(&sql_indexer_context.db)
+        .one(&httpd_context.db)
         .await
         .expect("Can't fetch blocks");
     assert_that!(block).is_some();
@@ -280,9 +255,9 @@ async fn no_sql_index_error_after_restart() {
     // 5. Send a transaction
     suite
         .send_message_with_gas(
-            &mut accounts["sender"],
-            2000,
-            Message::transfer(to, Coins::one(denom.clone(), 2_000).unwrap()).unwrap(),
+            &mut accounts.user1,
+            1_000_000,
+            Message::transfer(to, Coins::one(usdc::DENOM.clone(), 100).unwrap()).unwrap(),
         )
         .await
         .should_succeed();
@@ -298,7 +273,7 @@ async fn no_sql_index_error_after_restart() {
     // 6. Verify the block height 2 is indexed
     let block = entity::blocks::Entity::find()
         .order_by_desc(entity::blocks::Column::BlockHeight)
-        .one(&sql_indexer_context.db)
+        .one(&httpd_context.db)
         .await
         .expect("Can't fetch blocks");
     assert_that!(block).is_some();
@@ -312,13 +287,8 @@ async fn no_sql_index_error_after_restart() {
 /// Ensure that flatten events are indexed correctly.
 #[tokio::test(flavor = "multi_thread")]
 async fn index_block_events() {
-    let (indexer, indexer_context, _) = create_hooked_indexer().await;
-
-    let (mut suite, mut accounts) = TestBuilder::new_with_indexer(indexer)
-        .add_account("owner", Coin::new("usdc", 100_000).unwrap())
-        .add_account("sender", Coins::new())
-        .set_owner("owner")
-        .build();
+    let (mut suite, mut accounts, _, _, _, httpd_context, _, _, _db_guard) =
+        setup_test_naive_with_indexer(TestOption::default()).await;
 
     let replier_code = ContractBuilder::new(Box::new(replier::instantiate))
         .with_execute(Box::new(replier::execute))
@@ -328,7 +298,7 @@ async fn index_block_events() {
 
     let replier_addr = suite
         .upload_and_instantiate(
-            &mut accounts["owner"],
+            &mut accounts.owner,
             replier_code,
             &Empty {},
             "salt",
@@ -347,7 +317,7 @@ async fn index_block_events() {
     );
 
     suite
-        .execute(&mut accounts["owner"], replier_addr, &msg, Coins::default())
+        .execute(&mut accounts.owner, replier_addr, &msg, Coins::default())
         .await
         .should_succeed();
 
@@ -361,27 +331,27 @@ async fn index_block_events() {
 
     // ensure block was saved
     let block = entity::blocks::Entity::find()
-        .one(&indexer_context.db)
+        .one(&httpd_context.db)
         .await
         .expect("Can't fetch blocks");
     assert_that!(block).is_some();
     assert_that!(block.unwrap().block_height).is_equal_to(1);
 
     let transactions = entity::transactions::Entity::find()
-        .all(&indexer_context.db)
+        .all(&httpd_context.db)
         .await
         .expect("Can't fetch transactions");
     assert_that!(transactions).is_not_empty();
 
     let messages = entity::messages::Entity::find()
-        .all(&indexer_context.db)
+        .all(&httpd_context.db)
         .await
         .expect("Can't fetch messages");
     assert_that!(messages).is_not_empty();
 
     let events = entity::events::Entity::find()
         .filter(entity::events::Column::BlockHeight.eq(2))
-        .all(&indexer_context.db)
+        .all(&httpd_context.db)
         .await
         .expect("Can't fetch events");
     assert_that!(events).is_not_empty();
@@ -395,7 +365,7 @@ async fn index_block_events() {
     // check for parent events
     let events = entity::events::Entity::find()
         .filter(entity::events::Column::ParentId.is_not_null())
-        .all(&indexer_context.db)
+        .all(&httpd_context.db)
         .await
         .expect("Can't fetch events");
     assert_that!(events).is_not_empty();
