+++ dango/testing/tests/indexing/perps_events.rs
@@ -12,7 +12,7 @@ use {
 
 #[tokio::test(flavor = "multi_thread")]
 async fn index_perps_events() -> anyhow::Result<()> {
-    let (mut suite, mut accounts, _, contracts, _, dango_context, _, _db_guard) =
+    let (mut suite, mut accounts, _, contracts, _, dango_context, _, _, _db_guard) =
         setup_test_with_indexer(TestOption::default()).await;
 
     setup_perps_env(&mut suite, &mut accounts, &contracts, 2_000, 100_000).await;
