+++ dango/testing/tests/indexing/users.rs
@@ -11,7 +11,7 @@ use {
 
 #[tokio::test(flavor = "multi_thread")]
 async fn index_single_user_multiple_public_keys() -> anyhow::Result<()> {
-    let (suite, mut accounts, codes, contracts, validator_sets, dango_context, _, _db_guard) =
+    let (suite, mut accounts, codes, contracts, validator_sets, dango_context, _, _, _db_guard) =
         setup_test_with_indexer(TestOption::default()).await;
     let mut suite = HyperlaneTestSuite::new(suite, validator_sets, &contracts);
